refactor: remove typography barrel file
