+++ @udir-design/react/src/components/beta.ts
@@ -35,4 +35,7 @@ export * from './table';
 export * from './textfield/Textfield';
 export * from './toggleGroup/ToggleGroup';
 export * from './tooltip/Tooltip';
-export * from './typography';
+export * from './typography/heading/Heading';
+export * from './typography/label/Label';
+export * from './typography/paragraph/Paragraph';
+export * from './typography/validationMessage/ValidationMessage';
