--- @udir-design/react/src/components/typography/index.tsx
@@ -1,4 +1,0 @@
-export * from './heading/Heading';
-export * from './label/Label';
-export * from './paragraph/Paragraph';
-export * from './validationMessage/ValidationMessage';
